test: add e2e for localization
