+++ lib/i18n/client.js
@@ -5,11 +5,11 @@ import LanguageDetector from "i18next-browser-languagedetector";
 import {
   defaultNS,
   languages,
-  getOptions,
   namespaces,
   cookieName,
   fallbackLng,
 } from "./settings";
+import { getOptions } from "./options";
 import { loadResources } from "./index";
 
 const runsOnServerSide = typeof window === "undefined";
