+++ lib/i18n/index.ts
@@ -1,7 +1,8 @@
 import { createInstance } from "i18next";
 import resourcesToBackend from "i18next-resources-to-backend";
 import { initReactI18next } from "react-i18next/initReactI18next";
-import { fallbackLng, getOptions } from "./settings";
+import { fallbackLng } from "./settings";
+import { getOptions } from "./options";
 
 export const loadResources = resourcesToBackend(
   (language: string, namespace: string, callback) => {
